rename config to settings for clarity
